almost working disassembler for ALU instructions
