+++ Cargo.toml
@@ -6,5 +6,5 @@ members = [
     "macros",
     "backends/common",
     "backends/riscv",
-    "backends/riscv/fuzz",
+    "fuzz",
 ]
