+++ backends/riscv/src/utils/mod.rs
@@ -1,0 +1,3 @@
+mod instr;
+
+pub use instr::*;
