+++ codecov.yaml
@@ -1,2 +1,4 @@
 ignore:
   - "macros"
+  - "fuzz"
+  - "**/tests/**"
