+++ core/src/error.rs
@@ -2,6 +2,7 @@ pub enum Error {
     FileError,
     ElfParseError,
     NoProgramSegments,
+    Unknown,
 }
 
 pub type Result<T> = std::result::Result<T, Error>;
