+++ core/src/operation.rs
@@ -159,6 +159,10 @@ impl Operation {
     pub fn get_impl(&self) -> Rc<RefCell<OperationImpl>> {
         self.r#impl.clone()
     }
+
+    pub fn add_operand(&mut self, operand: Operand) {
+        self.r#impl.borrow_mut().operands.push(operand);
+    }
 }
 
 pub trait Op {
